FIX (UI): Made Canvas a Prefab
Made the restart canvas a prefab to be instantiated rather than a game object in the scene to be turned on and off. This will make editing the canvas in the future easier.

The text of this canvas is now smaller and appears in the top left instead of the center.